// src/interfaces/DataInterface.ts
export interface DataItem {
  zipcode: string;
  latitude: number | undefined;
  longitude: number | undefined
  city: string | undefined
  state_id: string | undefined
  state_name: string | undefined
  density: number | undefined
  county_fips: number | undefined
  county_name: string | undefined
  timezone: string | undefined
  hispanos_in_us: number | undefined
  population: number | undefined
}